remve tenants from profiles worker
