+++ services/apps/profiles_worker/src/activities/common.ts
@@ -1,11 +1,9 @@
 import {
   getAffiliationsLastCheckedAt,
   getAllMemberIdsPaginated,
-  getAllTenants as getAllTenantsFromDAL,
   getMemberIdsWithRecentRoleChanges,
   updateAffiliationsLastCheckedAt,
 } from '@crowd/data-access-layer/src/old/apps/profiles_worker'
-import { ITenant } from '@crowd/types'
 
 import { svc } from '../main'
 
@@ -14,18 +12,16 @@ export async function getAffiliationsLastCheckedAtOfTenant(tenantId: string): Pr
 }
 
 export async function getMemberIdsForAffiliationUpdates(
-  tenantId: string,
   affiliationsLastChecked: string,
   limit: number,
   offset: number,
 ): Promise<string[]> {
   if (!affiliationsLastChecked) {
-    return getAllMemberIdsPaginated(svc.postgres.writer, tenantId, limit, offset)
+    return getAllMemberIdsPaginated(svc.postgres.writer, limit, offset)
   }
 
   return getMemberIdsWithRecentRoleChanges(
     svc.postgres.writer,
-    tenantId,
     affiliationsLastChecked,
     limit,
     offset,
@@ -34,8 +30,4 @@ export async function getMemberIdsForAffiliationUpdates(
 
 export async function updateAffiliationsLastCheckedAtOfTenant(tenantId: string): Promise<void> {
   await updateAffiliationsLastCheckedAt(svc.postgres.writer, tenantId)
-}
-
-export async function getAllTenants(): Promise<ITenant[]> {
-  return getAllTenantsFromDAL(svc.postgres.writer)
 }
