+++ services/apps/profiles_worker/src/types/member.ts
@@ -8,7 +8,4 @@ export interface MemberUpdateInput {
 
 export interface IRecalculateAffiliationsForNewRolesInput {
   offset?: number
-  tenant: {
-    id: string
-  }
 }
