+++ services/apps/profiles_worker/src/types/organization.ts
@@ -1,5 +1,4 @@
 export interface IOrganizationProfileSyncInput {
-  tenantId: string
   organization: {
     id: string
   }
