+++ services/apps/profiles_worker/src/workflows.ts
@@ -1,11 +1,11 @@
 import { memberUpdate } from './workflows/member/memberUpdate'
 import { recalculateAffiliationsForNewRoles } from './workflows/member/recalculateAffiliationsForNewRoles'
-import { triggerRecalculateAffiliationsForEachTenant } from './workflows/member/triggerRecalculateAffiliationsForEachTenant'
+import { triggerRecalculateAffiliations } from './workflows/member/triggerRecalculateAffiliations'
 import { organizationUpdate } from './workflows/organization/organizationUpdate'
 
 export {
   memberUpdate,
   organizationUpdate,
   recalculateAffiliationsForNewRoles,
-  triggerRecalculateAffiliationsForEachTenant,
+  triggerRecalculateAffiliations,
 }
