+++ services/apps/profiles_worker/src/workflows/member/recalculateAffiliationsForNewRoles.ts
@@ -6,6 +6,7 @@ import {
   proxyActivities,
 } from '@temporalio/workflow'
 
+import { getDefaultTenantId } from '@crowd/common'
 import { TemporalWorkflowId } from '@crowd/types'
 
 import * as activities from '../../activities'
@@ -29,11 +30,12 @@ const {
 export async function recalculateAffiliationsForNewRoles(
   input: IRecalculateAffiliationsForNewRolesInput,
 ): Promise<void> {
-  const affiliationsLastChecked = await getAffiliationsLastCheckedAtOfTenant(input.tenant.id)
+  const tenantId = getDefaultTenantId()
+  const affiliationsLastChecked = await getAffiliationsLastCheckedAtOfTenant(tenantId)
 
   console.log(
     `Recalculating affiliations for tenant ${
-      input.tenant.id
+      tenantId
     } with affiliationsLastChecked: ${affiliationsLastChecked} and offset ${input.offset || 0}`,
   )
 
@@ -41,14 +43,13 @@ export async function recalculateAffiliationsForNewRoles(
   const offset = input.offset || 0
 
   const memberIds = await getMemberIdsForAffiliationUpdates(
-    input.tenant.id,
     affiliationsLastChecked,
     MEMBER_PAGE_SIZE,
     offset,
   )
 
   if (memberIds.length === 0) {
-    await updateAffiliationsLastCheckedAtOfTenant(input.tenant.id)
+    await updateAffiliationsLastCheckedAtOfTenant(tenantId)
     return
   }
 
@@ -57,7 +58,7 @@ export async function recalculateAffiliationsForNewRoles(
   await Promise.all(
     memberIds.map((id) => {
       return executeChild(memberUpdate, {
-        workflowId: `${TemporalWorkflowId.MEMBER_UPDATE}/${input.tenant.id}/${id}`,
+        workflowId: `${TemporalWorkflowId.MEMBER_UPDATE}/${id}`,
         cancellationType: ChildWorkflowCancellationType.ABANDON,
         parentClosePolicy: ParentClosePolicy.PARENT_CLOSE_POLICY_ABANDON,
         retry: {
@@ -72,15 +73,12 @@ export async function recalculateAffiliationsForNewRoles(
             },
           },
         ],
-        searchAttributes: {
-          TenantId: [input.tenant.id],
-        },
+        searchAttributes: {},
       })
     }),
   )
 
   await continueAsNew<typeof recalculateAffiliationsForNewRoles>({
-    tenant: input.tenant,
     offset: offset + MEMBER_PAGE_SIZE,
   })
 }
