+++ services/libs/data-access-layer/src/old/apps/profiles_worker/index.ts
@@ -3,7 +3,6 @@ import QueryStream from 'pg-query-stream'
 
 import { DbConnOrTx, DbStore } from '@crowd/database'
 import { getServiceChildLogger } from '@crowd/logging'
-import { ITenant } from '@crowd/types'
 
 import { insertActivities } from '../../../activities'
 import { findMemberAffiliations } from '../../../member_segment_affiliations'
@@ -197,22 +196,15 @@ export async function getAffiliationsLastCheckedAt(db: DbStore, tenantId: string
   }
 }
 
-export async function getAllMemberIdsPaginated(
-  db: DbStore,
-  tenantId: string,
-  limit: number,
-  offset: number,
-) {
+export async function getAllMemberIdsPaginated(db: DbStore, limit: number, offset: number) {
   try {
     const results: IMemberId[] = await db.connection().any(
       `
       select id from members
-      where "tenantId" = $(tenantId)
       order by id asc
       limit $(limit)
       offset $(offset);`,
       {
-        tenantId,
         limit,
         offset,
       },
@@ -225,7 +217,6 @@ export async function getAllMemberIdsPaginated(
 
 export async function getMemberIdsWithRecentRoleChanges(
   db: DbStore,
-  tenantId: string,
   affiliationsLastChecked: string,
   limit: number,
   offset: number,
@@ -236,8 +227,7 @@ export async function getMemberIdsWithRecentRoleChanges(
       select distinct mo."memberId" as id from "memberOrganizations" mo
       join "members" m on mo."memberId" = m."id"
       where
-            m."tenantId" = $(tenantId)
-            and (
+            (
             mo."createdAt" > $(affiliationsLastChecked) or
             mo."updatedAt" > $(affiliationsLastChecked) or
             mo."deletedAt" > $(affiliationsLastChecked)
@@ -247,7 +237,6 @@ export async function getMemberIdsWithRecentRoleChanges(
       offset $(offset);`,
 
       {
-        tenantId,
         affiliationsLastChecked,
         limit,
         offset,
@@ -276,25 +265,4 @@ export async function updateAffiliationsLastCheckedAt(
   } catch (err) {
     throw new Error(err)
   }
-}
-
-export async function getAllTenants(db: DbStore): Promise<ITenant[]> {
-  let rows: ITenant[] = []
-  try {
-    rows = await db.connection().query(`
-      select
-        id as "tenantId",
-        plan
-      from tenants
-      where "deletedAt" is null
-        and plan IN ('Scale', 'Growth', 'Essential', 'Enterprise')
-        and ("trialEndsAt" > NOW() or "trialEndsAt" is null);
-    `)
-  } catch (err) {
-    this.log.error('Error while getting all tenants', err)
-
-    throw new Error(err)
-  }
-
-  return rows
 }
